Collapse multiple results with the same matches in a common prefix
When multiple results have a common prefix and the matches are all
contained within the prefix, return only one result for the prefix
rather than one for each of the matches.

eg. Given:

 /home/user/foo
 /home/user/bar
 /home/user/baz

A search for 'user' will return:

 /home/user

The ID assignment has also been changed to reset the IDs after each
query to reduce the amount of typing needed to navigate to results
from the match list.